PreviewLinkButton component
